requirements.txt
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
comm==0.2.2
cssselect2==0.7.0
curl_cffi==0.12.0
dateutils==0.6.12
debugpy==1.8.6
decorator==5.1.1
dnspython==2.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.56.0
frozendict==2.4.6
h11==0.14.0
html2image==2.0.5
html5lib==1.1
idna==3.10
imgkit==1.2.3
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multitasking==0.0.11
nest-asyncio==1.6.0
nsepython==2.94
nsetools==1.0.11
numpy==2.1.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfkit==1.0.0
peewee==3.17.8
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==6.31.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
pyee==11.1.1
Pygments==2.18.0
pymongo==4.10.1
pyphen==0.17.2
pyppeteer
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
reportlab==4.3.1
requests==2.32.3
scipy==1.15.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==1.26.20
wcwidth==0.2.13
weasyprint==64.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets
wsproto==1.2.0
xlrd==2.0.1
yfinance
zipp==3.21.0
zopfli==0.2.3.post1
